Moving service to services folder
